+++ src/components/ListSkeletons/ActivitySkeleton.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const ActivitySkeleton = () => {
+	return (
+		<div className="w-11/12 py-4">
+			<div className="skeleton w-full h-20 bg-slate-900 mx-4 rounded-none mt-4 lg:mt-0"></div>
+		</div>
+	);
+};
+
+export default ActivitySkeleton;
